+++ src/bootstrap.js
@@ -7,6 +7,7 @@ async function startup({rootURI}) {
   const scope = {Zotero, ChromeUtils, Services, URL: Zotero.getMainWindow().URL};
   Services.scriptloader.loadSubScript(rootURI + 'note-document.js', scope);
   Services.scriptloader.loadSubScript(rootURI + 'note-tabs.js', scope);
+  Services.scriptloader.loadSubScript(rootURI + 'codex-translation.js', scope);
   Services.scriptloader.loadSubScript(rootURI + 'connector.js', scope);
   connector = scope.createBridge();
   Zotero.ObsidianConnector = connector;